implement forward_iterator
